refactor(Lotto): 패키지 변경
변경 사항
- 클래스 위치를 lotto/back/domain 패키지로 변경
