+++ src/test/java/lotto/back/domain/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.back.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
